+++ packages/common/src/ui/components/domain/Dashboard/StatsPanel.tsx
