+++ packages/common/src/ui/components/domain/Dashboard/index.ts
@@ -1,0 +1,2 @@
+export * from './StatsPanel';
+export * from './Widget';
