+++ packages/common/src/ui/components/domain/index.ts
@@ -1,1 +1,2 @@
+export * from './Dashboard';
 export * from './StatusCrumbs';
