+++ packages/dashboard/src/widgets/index.ts
@@ -1,3 +1,1 @@
-export * from './InboundShipmentsWidget';
-export * from './OutboundShipmentsWidget';
 export * from './StockWidget';
