+++ packages/invoices/src/InboundShipment/Dashboard/index.ts
@@ -1,0 +1,1 @@
+export * from './Widget';
