+++ packages/invoices/src/OutboundShipment/Dashboard/index.ts
@@ -1,0 +1,1 @@
+export * from './Widget';
