+++ packages/invoices/src/index.ts
@@ -1,1 +1,3 @@
 export { default as InvoiceService } from './InvoiceService';
+export { OutboundShipmentWidget } from './OutboundShipment/Dashboard';
+export { InboundShipmentWidget } from './InboundShipment/Dashboard';
